Remove foreign key that causing, other plugin dependencies unable to refresh its database tables.
- Add conditional data seeder, when the current active theme code is flametree-theme seeder will autoload.
- Fix broken link in products management page

+++ essentials/updates/update_blog_plugin_table.php
@@ -9,25 +9,31 @@ class UpdateBlogPlugin extends Migration
 {
     public function up()
     {
-        if (Schema::hasTable('rainlab_blog_categories')) {
-            Schema::table('rainlab_blog_categories', function (Blueprint $table) {
-                $table->string('featured_image')->nullable()->after('description');
-                $table->unsignedInteger('user_group')->nullable()->after('id');
-            });
+        if (PluginManager::instance()->hasPlugin('RainLab.Blog')) {
+
+            if (Schema::hasTable('rainlab_blog_categories')) {
+                Schema::table('rainlab_blog_categories', function (Blueprint $table) {
+                    $table->string('featured_image')->nullable()->after('description');
+                    $table->unsignedInteger('user_group')->nullable()->after('id');
+                });
+            }
         }
     }
 
     public function down()
     {
-        if(Schema::hasTable('rainlab_blog_categories')){
-            Schema::table('rainlab_blog_categories', function (Blueprint $table) {
-                if (Schema::hasColumn('rainlab_blog_categories', 'featured_image')) {
-                    $table->dropColumn('featured_image');
-                }
-                if (Schema::hasColumn('rainlab_blog_categories', 'user_group')) {
-                    $table->dropColumn('user_group');
-                }
-            });
+        if (PluginManager::instance()->hasPlugin('RainLab.Blog')) {
+
+            if (Schema::hasTable('rainlab_blog_categories')) {
+                Schema::table('rainlab_blog_categories', function (Blueprint $table) {
+                    if (Schema::hasColumn('rainlab_blog_categories', 'featured_image')) {
+                        $table->dropColumn('featured_image');
+                    }
+                    if (Schema::hasColumn('rainlab_blog_categories', 'user_group')) {
+                        $table->dropColumn('user_group');
+                    }
+                });
+            }
         }
 
     }
